refactor(126318): :hammer: refatora formulario de alteracao de cardapio pt6
