+++ src/components/AlteracaoDeCardapio/Escola/components/ModalConfirmaAlteracao/index.jsx
@@ -1,7 +1,10 @@
+import { Botao } from "components//Shareable/Botao";
+import {
+  BUTTON_STYLE,
+  BUTTON_TYPE,
+} from "components/Shareable/Botao/constants";
 import React from "react";
 import { Modal } from "react-bootstrap";
-import Botao from "../../Shareable/Botao";
-import { BUTTON_TYPE, BUTTON_STYLE } from "../../Shareable/Botao/constants";
 
 export default (props) => (
   <Modal show={props.showModal} onHide={props.closeModal}>
